requirements.txt
Django
djangorestframework
psycopg2
django-rest-knox
pillow